requirements.txt
# Web 应用依赖
Flask>=2.3.0
Pillow>=9.0.0
Werkzeug>=2.3.0

# 生产环境部署（可选）
gunicorn>=21.0.0
